代码设置一些参数的示例
