+++ app/src/main/java/fanjh/mine/pullrefreshlayout/activity/ContentFixedActivity.java
@@ -1,15 +1,14 @@
-package fanjh.mine.pullrefreshlayout;
+package fanjh.mine.pullrefreshlayout.activity;
 
 import android.app.Activity;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.annotation.Nullable;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
-import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -17,31 +16,32 @@ import fanjh.mine.pulllayout.ILoadMoreListener;
 import fanjh.mine.pulllayout.IRefreshListener;
 import fanjh.mine.pulllayout.PullLayout;
 import fanjh.mine.pulllayout.PullLayoutOption;
+import fanjh.mine.pullrefreshlayout.R;
 
 /**
+ * 内容固定模式，类似于swipeRefresh
  * Created by xuws on 2017/5/11.
  */
 
-public class ListViewDemoActivity extends Activity {
+public class ContentFixedActivity extends Activity {
 
-    ListViewDemoActivity mContext;
-    ListView contentListView;
+    ContentFixedActivity mContext;
+    RecyclerView contentRecycleView;
     PullLayout pullLayout;
-
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_content_fixed);
         mContext = this;
-        setContentView(R.layout.activity_content_listview);
-        initView();
+        initVeiw();
     }
 
-    private void initView(){
-        pullLayout = (PullLayout) findViewById(R.id.listview_pullLayout);
-        contentListView = (ListView) findViewById(R.id.content_listview);
+    private void initVeiw(){
+        pullLayout = (PullLayout) findViewById(R.id.content_fixed_pullLayout);
+        contentRecycleView = (RecyclerView) findViewById(R.id.content_fixedrecycleview);
+        contentRecycleView.setLayoutManager(new LinearLayoutManager(mContext));
         MyAdapter myAdapter = new MyAdapter();
-        contentListView.setAdapter(myAdapter);
-
+        contentRecycleView.setAdapter(myAdapter);
         /**
          * 刷新滑动监听
          */
@@ -111,53 +111,46 @@ public class ListViewDemoActivity extends Activity {
             @Override
             public boolean canUpTpDown() {
                 //recycleView不能向上滑动，即滑到底了，这时候，可以加载更多
-                return !contentListView.canScrollVertically(-1);
+                return !contentRecycleView.canScrollVertically(-1);
             }
 
             @Override
             public boolean canDownToUp() {
                 //recycleview不能向下滑动，即在首部了，这时候，可以下拉刷新
-                return !contentListView.canScrollVertically(1);
+                return !contentRecycleView.canScrollVertically(1);
             }
         });
     }
 
-    class MyAdapter extends BaseAdapter{
+    class MyAdapter extends RecyclerView.Adapter{
 
         @Override
-        public int getCount() {
+        public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+            TextView textView = new TextView(mContext);
+            textView.setTextSize(16);
+            textView.setGravity(Gravity.CENTER);
+            textView.setPadding(10,10,10,10);
+            return new ViewHolder(textView);
+        }
+
+        @Override
+        public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
+             ViewHolder viewHolder = (ViewHolder) holder;
+             viewHolder.textView.setText("item" + position);;
+        }
+
+        @Override
+        public int getItemCount() {
             return 50;
         }
 
-        @Override
-        public Object getItem(int position) {
-            return null;
-        }
+        class ViewHolder extends RecyclerView.ViewHolder {
 
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            ViewHolder holder = null;
-            if (convertView == null) {
-                convertView = LayoutInflater.from(mContext).inflate(R.layout.item_listview, null, false);
-                holder = new ViewHolder(convertView);
-                convertView.setTag(holder);
-            } else {
-                holder = (ViewHolder) convertView.getTag();
+            TextView textView;
+            public ViewHolder(View itemView) {
+                super(itemView);
+                textView = (TextView) itemView;
             }
-
-            holder.textView.setText("item" + position);
-            return convertView;
         }
-
-        class ViewHolder{
-           TextView textView;
-            public ViewHolder(View view){
-                textView = (TextView) view.findViewById(R.id.tv_textView);
-            }}
     }
 }
