+++ app/src/main/java/fanjh/mine/pullrefreshlayout/footerview/FooterProgressView.java
@@ -1,8 +1,10 @@
-package fanjh.mine.pullrefreshlayout;
+package fanjh.mine.pullrefreshlayout.footerview;
 
 import android.content.Context;
 import android.util.AttributeSet;
+import android.view.Gravity;
 import android.widget.ProgressBar;
+import android.widget.TextView;
 
 import fanjh.mine.pulllayout.ILoadMoreListener;
 
@@ -11,14 +13,14 @@ import fanjh.mine.pulllayout.ILoadMoreListener;
 * @date 2017/5/4 11:10
 * @description 底部视图
 **/
-public class HeaderProgressView extends ProgressBar implements ILoadMoreListener{
+public class FooterProgressView extends ProgressBar implements ILoadMoreListener{
     private boolean isLoadMoreing;
 
-    public HeaderProgressView(Context context) {
+    public FooterProgressView(Context context) {
         this(context,null);
     }
 
-    public HeaderProgressView(Context context, AttributeSet attrs) {
+    public FooterProgressView(Context context, AttributeSet attrs) {
         super(context, attrs);
         setIndeterminate(true);
     }
