+++ app/src/main/res/layout/activity_content_listview.xml
@@ -5,11 +5,6 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    app:maxDownOffset="240dp"
-    app:maxUpOffset="120dp"
-    app:refreshOffset="60dp"
-    app:loadMoreOffset="60dp"
-    app:contentFixed="false"
     app:disabledNestedScrolling="false">
 
     <ListView
@@ -17,12 +12,12 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-    <fanjh.mine.pullrefreshlayout.HeaderProgressView
+    <fanjh.mine.pullrefreshlayout.headerview.HeaderProgressView
         android:id="@+id/listview_headerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
 
-    <fanjh.mine.pullrefreshlayout.FooterProgressView
+    <fanjh.mine.pullrefreshlayout.footerview.FooterProgressView
         android:id="@+id/listview_footerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
