+++ app/src/main/res/layout/activity_content_nestscrolling.xml
@@ -68,12 +68,12 @@
 
     </LinearLayout>
 
-    <fanjh.mine.pullrefreshlayout.HeaderProgressView
+    <fanjh.mine.pullrefreshlayout.headerview.HeaderProgressView
         android:id="@+id/nestscrolling_headerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
 
-    <fanjh.mine.pullrefreshlayout.FooterProgressView
+    <fanjh.mine.pullrefreshlayout.footerview.FooterProgressView
         android:id="@+id/nestscrolling_footerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
