+++ app/src/main/res/layout/activity_content_scrollview.xml
@@ -55,12 +55,12 @@
     </ScrollView>
 
 
-    <fanjh.mine.pullrefreshlayout.HeaderProgressView
+    <fanjh.mine.pullrefreshlayout.headerview.HeaderProgressView
         android:id="@+id/scrollview_headerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
 
-    <fanjh.mine.pullrefreshlayout.FooterProgressView
+    <fanjh.mine.pullrefreshlayout.footerview.FooterProgressView
         android:id="@+id/scrollview_footerview"
         android:layout_width="match_parent"
         android:layout_height="60dp" />
