+++ app/src/main/res/layout/activity_main.xml
@@ -51,12 +51,12 @@
 
     </android.support.v4.widget.NestedScrollView>
 
-    <fanjh.mine.pullrefreshlayout.HeaderView
+    <fanjh.mine.pullrefreshlayout.headerview.HeaderView
         android:id="@+id/hv_header"
         android:layout_width="match_parent"
         android:layout_height="60dp"/>
 
-    <fanjh.mine.pullrefreshlayout.FooterView
+    <fanjh.mine.pullrefreshlayout.footerview.FooterView
         android:id="@+id/hv_footer"
         android:layout_width="match_parent"
         android:layout_height="60dp"/>
